requirements.txt
# Core Flask
flask
python-dotenv
flask-sqlalchemy
flask-migrate

# Database Drivers
psycopg2-binary

# Security & Utilities
flask-cors
flask-jwt-extended
gunicorn
passlib[bcrypt]

# AI & Chatbot
openai